Crate fastcrypto_tbls

An experimental crate that implements threshold BLS (tBLS) and distributed key generation (DKG) protocols.

Modules

ecies

mocked_dkg

This module contains a set of functions for emulating the outputs of a DKG protocol. Those outputs can be used as inputs to the threshold-BLS functions.

polynomial

random_oracle

tbls

types