Struct fastcrypto_tbls::dkg::Party

pub struct Party<G: GroupElement, EG: GroupElement> { /* private fields */ }

Party in the DKG protocol.

Implementations

impl<G: GroupElement, EG> Party<G, EG>where <G as GroupElement>::ScalarType: Serialize + DeserializeOwned, EG: Serialize + GroupElement, <EG as GroupElement>::ScalarType: HashToGroupElement,

A dealer in the DKG ceremony.

Can be instantiated with G1Curve or G2Curve.

pub fn new<R: AllowedRng>( ecies_sk: PrivateKey<EG>, nodes: Nodes<EG>, threshold: u32, random_oracle: RandomOracle, rng: &mut R ) -> Result<Self, FastCryptoError>

Create a new ECIES private key and send the public key to all parties.

After all parties have sent their ECIES public keys, create the set of nodes.

Create a new Party instance with the ECIES private key and the set of nodes.

pub fn threshold(&self) -> u32

pub fn create_first_message<R: AllowedRng>( &self, rng: &mut R ) -> FirstMessage<G, EG>

Create the first message to be broadcasted.

pub fn create_second_message<R: AllowedRng>( &self, messages: &[FirstMessage<G, EG>], rng: &mut R ) -> Result<(SharesMap<G>, SecondMessage<EG>), FastCryptoError>

Process the first messages of exactly ‘threshold’ nodes and create the second message to be broadcasted. The second message contains the list of complaints on invalid shares. In addition, it returns a set of valid shares (so far). Since we assume that at most t-1 of the nodes are malicious, we only need messages from t nodes to guarantee an unbiasable and unpredictable beacon. (The result is secure with rushing adversaries according to https://eprint.iacr.org/2021/005.pdf.)

pub fn process_responses( &self, first_messages: &[FirstMessage<G, EG>], second_messages: &[SecondMessage<EG>], shares: SharesMap<G>, minimal_threshold: usize ) -> Result<SharesMap<G>, FastCryptoError>

Process all the second messages, check all complaints, and update the local set of valid shares accordingly.

minimal_threshold is the minimal number of second round messages we expect. Its value is application dependent but in most cases it should be at least 2t-1 to guarantee that at least t honest nodes have valid shares.

pub fn aggregate( &self, first_messages: &[FirstMessage<G, EG>], shares: SharesMap<G> ) -> DkgOutput<G, EG>

Aggregate the valid shares (as returned from the previous step) and the public key.