Module fastcrypto_tbls::mocked_dkg

This module contains a set of functions for emulating the outputs of a DKG protocol. Those outputs can be used as inputs to the threshold-BLS functions.

These functions are NOT SECURE and intended to be used only as an interim alternative until the DKG protocol is implemented.

use std::num::NonZeroU32;
use fastcrypto_tbls::mocked_dkg;
use fastcrypto_tbls::types::ThresholdBls12381MinSig;
use fastcrypto_tbls::tbls::ThresholdBls;

let threshold: u32 = 2;
let epoch: u64 = 10;
const MSG: [u8; 4] = [1, 2, 3, 4];

// Parties get their private shares from the mock.
let (share1, _, _) = mocked_dkg::generate_share_and_public_keys(threshold, epoch, NonZeroU32::new(1).unwrap());
let (share2, _, _) = mocked_dkg::generate_share_and_public_keys(threshold, epoch, NonZeroU32::new(2).unwrap());

// Parties may sign messages with their shares.
let sig1 = ThresholdBls12381MinSig::partial_sign(&share1, &MSG);
let sig2 = ThresholdBls12381MinSig::partial_sign(&share2, &MSG);

// Anyone can check partial signatures.
let (bls_pk, vss_pk) = mocked_dkg::generate_public_keys(threshold, epoch);
assert!(ThresholdBls12381MinSig::partial_verify(&vss_pk, &MSG, &sig1).is_ok());
assert!(ThresholdBls12381MinSig::partial_verify(&vss_pk, &MSG, &sig2).is_ok());

// Anyone can aggregate the partial signatures and verify the full signature.
let sig = ThresholdBls12381MinSig::aggregate(threshold, &[sig1, sig2]).unwrap();
assert!(ThresholdBls12381MinSig::verify(&bls_pk, &MSG, &sig).is_ok());

Functions

generate_full_key_pair

Emulate the output of the previous key recovery protocol.

generate_public_keys

Emulate the public output of DKG.

generate_share_and_public_keys

Emulate the output of DKG for a given id.