fastcrypto_tbls

Module threshold_schnorr

This module implements threshold Schnorr signatures. The signatures are compatible with BIP-0340.

It provides the following protocols:

A Distributed Key Generation (DKG) protocol to generate a shared signing key without a trusted dealer. The protocol also allows resharing of a share from a previous DKG, allowing for key rotation. This is implemented in the avss module.

A protocol to generate a batch of secret shared nonces for signing. This is implemented in the batch_avss module.

A presigning protocol to create presigning tuples from the secret shared nonces. This is implemented in the presigning module. The presigning tuples can be created in advance of knowing the message to be signed, and one tuple is consumed for each signature.

A signing protocol which allows parties to create partial signatures from a presigning tuple and aggregate them into a full signature if there are enough partial signatures. This is implemented in the signing module.

For both the DKG and nonce generation protocols, it is assumed that each party has an encryption key pair (ECIES) and these public keys are known to all parties. These can be reused for all instances of the protocols.

The thresholds are defined as follows:

n = total number of parties

f = maximum number of Byzantine parties

t = threshold for signing

The following conditions must hold: t + 2f ≤ n and t > f.

Modules

avss

Implementation of an asynchronous verifiable secret sharing (AVSS) protocol to distribute shares for a secret to a set of receivers. A receiver can verify that the secret being shared is the same as a share from a previous round (e.g., the secret key share of a threshold signature).

batch_avss

Implementation of an asynchronous verifiable secret sharing (AVSS) protocol to distribute secret shares for a batch of random nonces. The size of the batch is proportional to the Dealer’s weight.

complaint

key_derivation

presigning

reed_solomon

signing

Type Aliases

Address

An address on the Sui network.

G

The group to use for the signing

S

Default scalar