fastcrypto_tbls::threshold_schnorr

Module batch_avss

Implementation of an asynchronous verifiable secret sharing (AVSS) protocol to distribute secret shares for a batch of random nonces. The size of the batch is proportional to the Dealer’s weight.

Before the protocol starts, the following setup is needed:

Each receiver has a encryption key pair (ECIES) and these public keys are known to all parties.

The public keys along with the weights of each receiver are known to all parties and defined in the Nodes structure.

Define a new Dealer with the secrets who begins by calling Dealer::create_message.

Structs

Dealer

This represents a Dealer in the AVSS. There is exactly one dealer who creates the shares and broadcasts the encrypted shares.

Message

The message broadcast by the dealer.

Receiver

This represents a Receiver in the AVSS who receives shares from the Dealer.

ReceiverOutput

The output of a receiver which is a batch of shares and public keys for all nonces.

ShareBatch

A batch of shares for a single share index, containing shares for each secret and one for the “blinding” polynomial.

SharesForNode

This represents a set of shares for a node. A total of L secrets/nonces are being shared, If we say that node i has a weight W_i, we have indices().len() == shares_for_secret(i).len() == weight() = W_i

Enums

ProcessedMessage

The result of processing a message by a receiver: either valid shares or a complaint.