fastcrypto_tbls::threshold_schnorr::signing

Function aggregate_signatures

pub fn aggregate_signatures(
    message: &[u8],
    public_presig: &G,
    beacon_value: &S,
    partial_signatures: &[Eval<S>],
    threshold: u16,
    verifying_key: &G,
    derivation_address: Option<&Address>,
) -> FastCryptoResult<SchnorrSignature>

Given enough partial signatures, aggregate them into a full signature and verify it. The signature produced follows the BIP-0340 standard.

If a derivation index is provided, a new verifying key is derived for this index (see [derive_verifying_key]), and the signature is adjusted accordingly. The signature will be valid for the derived verifying key.

Returns an InputTooShort error if not enough partial signatures are provided. GeneralOpaqueError is returned if the computed nonce R is the identity element. InvalidSignature is returned if the aggregated signature does not verify. InvalidInput is returned if the provided verifying key is the identity element.