fastcrypto_tbls::threshold_schnorr

Trait Certificate

pub trait Certificate {
    type Payload;

    // Required methods
    fn signers(&self) -> &BTreeSet<PartyId>;
    fn payload(&self) -> &Self::Payload;
    fn verify(&self) -> FastCryptoResult<()>;

    // Provided method
    fn into_verified(&self) -> FastCryptoResult<VerifiedCertificate<Self>>
       where Self: Clone { ... }
}

This represents a certificate over a payload that a subset of the parties have signed. Here, the implementation is abstract, and it is up to the caller to implement the actual verification functionality.

Required Associated Types

type Payload

Required Methods

fn signers(&self) -> &BTreeSet<PartyId>

fn payload(&self) -> &Self::Payload

fn verify(&self) -> FastCryptoResult<()>

Provided Methods

fn into_verified(&self) -> FastCryptoResult<VerifiedCertificate<Self>>
where Self: Clone,