Module fastcrypto_tbls::dkg

Structs

DkgOutput

DkgOutput is the final output of the DKG protocol in case it runs successfully. It can be used later with [ThresholdBls], see examples in tests.

EncryptedShare

EncryptedShare holds the ECIES encryption of a share destined to the receiver.

FirstMessage

[DkgFirstMessage] holds all encrypted shares a dealer sends during the first phase of the protocol.

Party

Party in the DKG protocol.

PkiNode

Generics below use G: GroupElement' for the group of the VSS public key, and EG: GroupElement’ for the group of the ECIES public key. PKI node, with a unique id and its encryption public key.

SecondMessage

A [DkgSecondMessage] is sent during the second phase of the protocol. It includes complaints created by receiver of invalid encrypted shares.

Enums

Complaint

A complaint/fraud claim against a dealer that created invalid encrypted share.

Type Definitions

Nodes

SharesMap

Mapping from node id to the share received from that leader.