fastcrypto_tbls

Module dkg_v1

Structs

Complaint

Assumptions:

Confirmation

A Confirmation is sent during the second phase of the protocol. It includes complaints created by receiver of invalid encrypted shares (if any).

Message

Message holds all encrypted shares a dealer sends during the first phase of the protocol.

Output

Output is the final output of the DKG protocol in case it runs successfully. It can be used later with [ThresholdBls], see examples in tests.

Party

Generics below use G: GroupElement' for the group of the VSS public key, and EG: GroupElement’ for the group of the ECIES public key. Party in the DKG protocol.

ProcessedMessage

Wrapper for collecting everything related to a processed message.

UsedProcessedMessages

Unique processed messages that are being used in the protocol.

VerifiedProcessedMessages

Processed messages that were not excluded after the third phase of the protocol.

Constants

DKG_MESSAGES_MAX_SIZE