Module fastcrypto_tbls::dl_verification

Functions

verify_deg_t_poly

Check that partial public keys form a polynomial of the right degree using the protocol of https://eprint.iacr.org/2017/216.pdf. deg_f should be n-k-2 if the polynomial is of degree k.

verify_equal_exponents

Checks if vectors v1=(a1G1, …, anG1) and v2=(a1’*G2, …, an’*G2) use ai = ai’ for all i, by computing <v1, e> and <v2, e> for a random e and checking if they are equal using pairing.

verify_pairs

Check that a pair (k, H) satisfies H = k*G using a random combination of the pairs and multi scalar multiplication.

verify_poly_evals

Checks that a given set of evaluations is consistent with a given polynomial in the exp by checking that (\sum r_i v_i)*G = \sum r_i p(i) for a random set of scalars r_i.

verify_triplets

Check that a triplet (k, G, H) satisfies H = k*G using a random combination of the triplets and multi scalar multiplication.