fastcrypto_tbls::dl_verification

Function verify_poly_evals

pub fn verify_poly_evals<G: GroupElement + MultiScalarMul, R: AllowedRng>(
    evals: &[Eval<G::ScalarType>],
    poly: &Poly<G>,
    rng: &mut R,
) -> FastCryptoResult<()>

Checks that a given set of evaluations is consistent with a given polynomial in the exp by checking that (\sum r_i v_i)*G = \sum r_i p(i) for a random set of scalars r_i.