Trait fastcrypto_tbls::tbls::ThresholdBls

pub trait ThresholdBls {
    type Private: Scalar;
    type Public: GroupElement<ScalarType = Self::Private>;
    type Signature: GroupElement<ScalarType = Self::Private> + HashToGroupElement;

    fn verify_pairings(
        pk: &Self::Public,
        sig: &Self::Signature,
        msg: &[u8]
    ) -> Result<(), FastCryptoError>;

    fn sign(private: &Self::Private, msg: &[u8]) -> Self::Signature { ... }
    fn verify(
        public: &Self::Public,
        msg: &[u8],
        sig: &Self::Signature
    ) -> Result<(), FastCryptoError> { ... }
    fn partial_sign(
        share: &Share<Self::Private>,
        msg: &[u8]
    ) -> PartialSignature<Self::Signature> { ... }
    fn partial_verify(
        vss_pk: &Poly<Self::Public>,
        msg: &[u8],
        partial_sig: &PartialSignature<Self::Signature>
    ) -> Result<(), FastCryptoError> { ... }
    fn aggregate(
        threshold: u32,
        partials: &[PartialSignature<Self::Signature>]
    ) -> Result<Self::Signature, FastCryptoError> { ... }
}

Trait ThresholdBls provides sign & verify functions for standard and partial BLS signatures.

Required Associated Types

type Private: Scalar

type Public: GroupElement<ScalarType = Self::Private>

Public represents the group over which the public keys are represented.

type Signature: GroupElement<ScalarType = Self::Private> + HashToGroupElement

Signature represents the group over which the signatures are represented.

Required Methods

fn verify_pairings(
pk: &Self::Public,
sig: &Self::Signature,
msg: &[u8]
) -> Result<(), FastCryptoError>

Curve dependent implementation of computing and comparing the pairings as part of the signature verification.

Provided Methods

fn sign(private: &Self::Private, msg: &[u8]) -> Self::Signature

Sign a message using the private key.

fn verify(
public: &Self::Public,
msg: &[u8],
sig: &Self::Signature
) -> Result<(), FastCryptoError>

Verify a signature on a given message.

fn partial_sign(
share: &Share<Self::Private>,
msg: &[u8]
) -> PartialSignature<Self::Signature>

Sign a message using the private share/partial key.

fn partial_verify(
vss_pk: &Poly<Self::Public>,
msg: &[u8],
partial_sig: &PartialSignature<Self::Signature>
) -> Result<(), FastCryptoError>

Verify a signature done by a partial key holder.

fn aggregate(
threshold: u32,
partials: &[PartialSignature<Self::Signature>]
) -> Result<Self::Signature, FastCryptoError>

Interpolate partial signatures to recover the full signature.