Trait fastcrypto_tbls::tbls::ThresholdBls

pub trait ThresholdBls {
    type Private: Scalar;
    type Public: GroupElement<ScalarType = Self::Private> + MultiScalarMul;
    type Signature: GroupElement<ScalarType = Self::Private> + HashToGroupElement + MultiScalarMul;

    // Required method
    fn verify(
        public: &Self::Public,
        msg: &[u8],
        sig: &Self::Signature
    ) -> FastCryptoResult<()>;

    // Provided methods
    fn partial_sign(
        share: &Share<Self::Private>,
        msg: &[u8]
    ) -> PartialSignature<Self::Signature> { ... }
    fn partial_sign_batch(
        shares: impl Iterator<Item = impl Borrow<Share<Self::Private>>>,
        msg: &[u8]
    ) -> Vec<PartialSignature<Self::Signature>> { ... }
    fn partial_verify(
        vss_pk: &Poly<Self::Public>,
        msg: &[u8],
        partial_sig: &PartialSignature<Self::Signature>
    ) -> FastCryptoResult<()> { ... }
    fn partial_verify_batch<R: AllowedRng>(
        vss_pk: &Poly<Self::Public>,
        msg: &[u8],
        partial_sigs: impl Iterator<Item = impl Borrow<PartialSignature<Self::Signature>>>,
        rng: &mut R
    ) -> FastCryptoResult<()> { ... }
    fn aggregate(
        threshold: u16,
        partials: impl Iterator<Item = impl Borrow<PartialSignature<Self::Signature>>> + Clone
    ) -> FastCryptoResult<Self::Signature> { ... }
}

Trait ThresholdBls provides sign & verify functions for standard and partial BLS signatures.

Required Associated Types

type Private: Scalar

type Public: GroupElement<ScalarType = Self::Private> + MultiScalarMul

Public represents the group over which the public keys are represented.

type Signature: GroupElement<ScalarType = Self::Private> + HashToGroupElement + MultiScalarMul

Signature represents the group over which the signatures are represented.

Required Methods

fn verify( public: &Self::Public, msg: &[u8], sig: &Self::Signature ) -> FastCryptoResult<()>

Verify a signature on a given message. This is standard BLS signature verification over the used curve construction.

Provided Methods

fn partial_sign( share: &Share<Self::Private>, msg: &[u8] ) -> PartialSignature<Self::Signature>

Sign a message using the private share/partial key.

fn partial_sign_batch( shares: impl Iterator<Item = impl Borrow<Share<Self::Private>>>, msg: &[u8] ) -> Vec<PartialSignature<Self::Signature>>

Sign a message using one of more private share/partial keys.

fn partial_verify( vss_pk: &Poly<Self::Public>, msg: &[u8], partial_sig: &PartialSignature<Self::Signature> ) -> FastCryptoResult<()>

Verify a signature done by a partial key holder.

fn partial_verify_batch<R: AllowedRng>( vss_pk: &Poly<Self::Public>, msg: &[u8], partial_sigs: impl Iterator<Item = impl Borrow<PartialSignature<Self::Signature>>>, rng: &mut R ) -> FastCryptoResult<()>

Verify a set of signatures done by a partial key holder. Randomly check if \sum r_i sig_i is a valid signature with public key \sum r_i p(i) G where r_i are random scalars, and p(i) are points on the polynomial.

fn aggregate( threshold: u16, partials: impl Iterator<Item = impl Borrow<PartialSignature<Self::Signature>>> + Clone ) -> FastCryptoResult<Self::Signature>

Interpolate partial signatures to recover the full signature. Uses the first threshold unique partials to interpolate the signature and the rest are ignored (thus if those partials are invalid, the resulting signature can still be valid).

Object Safety

This trait is not object safe.