fastcrypto_tbls::threshold_schnorr

Module avss

Implementation of an asynchronous verifiable secret sharing (AVSS) protocol to distribute shares for a secret to a set of receivers. A receiver can verify that the secret being shared is the same as a share from a previous round (e.g., the secret key share of a threshold signature).

Before the protocol starts, the following setup is needed:

Each receiver has an encryption key pair (ECIES) and these public keys are known to all parties.

The public keys along with the weights of each receiver are known to all parties and defined in the Nodes structure.

Define a new [crate::threshold_schnorr::Dealer] with the secrets, who begins by calling [crate::threshold_schnorr::Dealer::create_message].

Structs

Dealer

This represents a Dealer in the AVSS. There is exactly one dealer, who creates the shares and broadcasts the encrypted shares.

Message

The message broadcast by the dealer, containing the encrypted shares and the public keys of the nonces.

PartialOutput

The output of a receiver after a single instance of AVSS: The shares for each nonce + commitments for the next round.

Receiver

ReceiverOutput

The output after combining multiple PartialOutputs, either using [PartialOutput::complete_dkg] or [PartialOutput::complete_key_rotation].

SharesForNode

All the shares given to a node. One share per the node’s weight. These can be created either by decrypting the shares from the dealer (see Receiver::process_message) or by recovering them from complaint responses.

Enums

ProcessedMessage

The result of a Receiver processing a Message: Either valid shares or a complaint.