fastcrypto_tbls::threshold_schnorr::signing

Function generate_partial_signatures

pub fn generate_partial_signatures(
    message: &[u8],
    presignatures: &mut Presignatures,
    beacon_value: &S,
    my_signing_key_shares: &SharesForNode,
    verifying_key: &G,
    derivation_address: Option<&Address>,
) -> FastCryptoResult<(G, Vec<Eval<S>>)>

Generate partial threshold Schnorr signatures for a given message using a presigning triple. Returns also the public nonce R.

The signatures produced follow the BIP-0340 standard (https://github.com/bitcoin/bips/blob/master/bip-0340.mediawiki).

If a derivation index is provided, a new verifying key is derived for this index (see [derive_verifying_key]), and the signature is adjusted accordingly. The signature will be valid for the derived verifying key.

Returns an OutOfPresigs error if the presignatures iterator is exhausted. GeneralOpaqueError is returned if the generated nonce R is the identity element (should happen only with negligible probability). InvalidInput is returned if the verifying key is the identity element.