fastcrypto_tbls

Module types

Structs

IndexedValue

Wrapper of a value that is associated with a specific index.

ThresholdBls12381MinSig

Implementation of [ThresholdBls] for BLS12-381-min-sig. A variant for BLS12-381-min-pk can be defined in a similar way if needed in the future.

UnindexedValues

Basic wrapper of a set of values that are not associated with indexes, assuming the indexes are known to all parties. Used to reduce the size of the messages in the protocol.

Type Aliases

PrivateBlsKey

PublicBlsKey

PublicVssKey

RawSignature

Share

tBLS with ThresholdBls12381MinSig types.

ShareIndex

Indexes of shares/private keys (0 is reserved for the secret itself).

Signature