Module fastcrypto_tbls::types

Structs

IndexedValue

Wrapper of a value that is associated with a specific index.

ThresholdBls12381MinSig

Implementation of [ThresholdBls] for BLS12-381-min-sig. A variant for BLS12-381-min-pk can be defined in a similar way if needed in the future.

Type Definitions

EciesEncryption

PrivateBlsKey

PrivateEciesKey

ECIES related types with Ristretto points.

PublicBlsKey

PublicEciesKey

PublicVssKey

RawSignature

RecoveryPackage

Share

tBLS with ThresholdBls12381MinSig types.

ShareIndex

Indexes of shares/private keys (0 is reserved for the secret itself).

Signature